mapreduce 2/3/4 done+ 5/6/7 setup

+++ src/main/java/mapreduce/mapred4.java
@@ -88,7 +88,7 @@ public class mapred4 {
                 sum += inputvalue.get();
                 compteur++;
             }
-            int moyenne = (sum/compteur)/100;
+            double moyenne = ((double)sum/(double)compteur)/100.0;
             String smoyenne = String.valueOf(moyenne);
             System.out.println(moyenne);
             // create hbase put with rowkey as date
